+++ test/tokenizers_test.rb
@@ -5,9 +5,9 @@ class TokenizersTest < Minitest::Test
     tokenizer = Tokenizers.from_pretrained("bert-base-cased")
 
     # encode
-    encoded = tokenizer.encode("I can feel the magic, can you?", add_special_tokens: false)
-    expected_ids = [146, 1169, 1631, 1103, 3974, 117, 1169, 1128, 136]
-    expected_tokens = ["I", "can", "feel", "the", "magic", ",", "can", "you", "?"]
+    encoded = tokenizer.encode("I can feel the magic, can you?")
+    expected_ids = [101, 146, 1169, 1631, 1103, 3974, 117, 1169, 1128, 136, 102]
+    expected_tokens = ["[CLS]", "I", "can", "feel", "the", "magic", ",", "can", "you", "?", "[SEP]"]
     assert_equal expected_ids, encoded.ids
     assert_equal expected_tokens, encoded.tokens
 
@@ -19,7 +19,7 @@ class TokenizersTest < Minitest::Test
     tokenizer = Tokenizers.from_pretrained("gpt2")
 
     # encode
-    encoded = tokenizer.encode("Mythological creatures like the mighty gryphon inspire awe!", add_special_tokens: false)
+    encoded = tokenizer.encode("Mythological creatures like the mighty gryphon inspire awe!")
     expected_ids = [41444, 2770, 8109, 588, 262, 18680, 308, 563, 746, 261, 18330, 25030, 0]
     expected_tokens = ["Myth", "ological", "Ġcreatures", "Ġlike", "Ġthe", "Ġmighty", "Ġg", "ry", "ph", "on", "Ġinspire", "Ġawe", "!"]
     expected_word_ids = [0, 0, 1, 2, 3, 4, 5, 5, 5, 5, 6, 7, 8]
@@ -47,10 +47,17 @@ class TokenizersTest < Minitest::Test
   end
 
   def test_add_special_tokens
-     tokenizer = Tokenizers.from_pretrained("bert-base-cased")
-     encoded = tokenizer.encode("I can feel the magic, can you?", add_special_tokens: true)
-     expected = ["[CLS]", "I", "can", "feel", "the", "magic", ",", "can", "you", "?", "[SEP]"]
-     assert_equal expected, encoded.tokens
+    tokenizer = Tokenizers.from_pretrained("bert-base-cased")
+
+    # encode
+    encoded = tokenizer.encode("I can feel the magic, can you?", add_special_tokens: false)
+    expected_ids = [146, 1169, 1631, 1103, 3974, 117, 1169, 1128, 136]
+    expected_tokens = ["I", "can", "feel", "the", "magic", ",", "can", "you", "?"]
+    assert_equal expected_ids, encoded.ids
+    assert_equal expected_tokens, encoded.tokens
+
+    # decode
+    assert_equal "I can feel the magic, can you?", tokenizer.decode(encoded.ids)
   end
 
   def test_char_bpe_tokenizer
@@ -78,7 +85,7 @@ class TokenizersTest < Minitest::Test
 
   def test_multibyte_offsets
      tokenizer = Tokenizers.from_pretrained("gpt2")
-     encoded = tokenizer.encode("I wanted to convert 10000 ¥ to $.", add_special_tokens: true)
+     encoded = tokenizer.encode("I wanted to convert 10000 ¥ to $.")
      expected_tokens = ["I", "Ġwanted", "Ġto", "Ġconvert", "Ġ10000", "ĠÂ¥", "Ġto", "Ġ$", "."]
      expected_offsets = [[0, 1], [1, 8], [8, 11], [11, 19], [19, 25], [25, 27], [27, 30], [30, 32], [32, 33]]
 
@@ -90,7 +97,7 @@ class TokenizersTest < Minitest::Test
     tokenizer = Tokenizers.from_pretrained("bert-base-cased")
     question = "Am I allowed to pass two text arguments?"
     answer = "Yes I am!"
-    encoded = tokenizer.encode(question, answer, add_special_tokens: true)
+    encoded = tokenizer.encode(question, answer)
 
     expected_tokens = ["[CLS]", "Am", "I", "allowed", "to", "pass", "two", "text", "arguments", "?", "[SEP]", "Yes", "I", "am", "!", "[SEP]"]
     assert_equal expected_tokens, encoded.tokens